JBTM-3965 Allways enlist participants
